requirements.txt
python-dotenv>=1.0.0
websockets>=12.0
beautifulsoup4>=4.12.0
